+++ src/main/java/pl/kluczewski/currency_converter/model/CurrencyDto.java
@@ -3,12 +3,14 @@ package pl.kluczewski.currency_converter.model;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 
 @Builder
 @Getter
 public class CurrencyDto {
     String code;
-    double mid;
-    double result;
+    BigDecimal mid;
+    BigDecimal result;
     String effectiveDate;
 }
