+++ src/main/java/pl/kluczewski/currency_converter/webclient/dto/AllRatesDto.java
@@ -2,9 +2,11 @@ package pl.kluczewski.currency_converter.webclient.dto;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class AllRatesDto {
     private String currency;
     private String code;
-    private double mid;
+    private BigDecimal mid;
 }
